Update README.md
Task Title: Create a Responsive Website and Host it on Netlify 

Task Description: As a web developer, your task is to create a responsive website that works seamlessly on desktop, tablet, and mobile devices. Once the development is complete, you will host it on Netlify and share the link with the team. 

Task Steps: 

Gather all the necessary assets such as logos, images, fonts, and color schemes for the website. Link to The Resources: Webtaskresources- Get it from the mentioned source

Create a responsive design that works well on desktop, tablet, and mobile devices. Use grids and breakpoints to ensure that the design is optimized for different screen sizes. 

Code the website using React, HTML, CSS, and JavaScript. Ensure that the website is accessible and meets web standards. 

Use responsive design techniques such as media queries to ensure that the website is optimized for different screen sizes. 

Test the website thoroughly to ensure that it is bug-free and works well on different browsers and devices. 

Once the development is complete, host the website on Netlify.  

Share the link with the team in the corresponding platform that you received this task. Make any necessary changes and improvements to ensure that the website meets the project requirements. 

Resource for Task: Webtaskresources - Click on this link to view resources

 
figma Link : https://www.figma.com/file/Yuvgoqi2kww5U03XvqiE60/New-Web-Development-task?node-id=0%3A1&t=itjc1qsL4BaWliiS-1